Migration to Harmony2.0
